refactor(http): split job.rs into thin facade + sibling modules
Break the 746-line `crates/dcc-mcp-http/src/job.rs` module into focused
siblings using the `#[path]`-included submodule pattern established by
the file-logging / watcher / action-chain refactors:

- `job_types.rs` — `JobStatus`, `JobProgress`, `Job` (with the private
  `Job::new` constructor), `JobEvent`, `JobSubscriber`.
- `job_manager.rs` — the `JobManager` registry: creation, state
  transitions (`start` / `complete` / `fail` / `cancel` /
  `update_progress`), persistence write-through, subscriber emission,
  recovery, and GC helpers (`gc_stale`, `cleanup_older_than_hours`).
- `job_tests.rs` — the full lifecycle / cancellation / invalid-transition
  / GC / concurrent-create / serde unit tests.

`job.rs` becomes a thin facade that carries only the module-level
docs (concurrency model + state diagram + maintainer layout) and
`pub use` re-exports. No public API changes.

Documentation: `docs/guide/architecture.md` now records the new
layout under the `dcc-mcp-http` section.
